feat: add unit test framework and security hardening
Security improvements:
- Add rate limiting to distributed coordinator (DoS protection)
- Add JSON input sanitization to prevent injection attacks
- Add configurable bind address (instead of 0.0.0.0)
- Add TLS support placeholder for future encryption
- Replace exit() with error returns in SECP256K1 library
- Fix buffer overflow in wizard_client using fork/exec instead of system()
- Add is_safe_path() validation against shell metacharacters

Thread safety fixes:
- Make FINISHED_ITEMS atomic with std::atomic<uint64_t>
- Add thread-safe random number generation (thread_rand)

Unit test framework (69 tests):
- test_int.cpp: 41 tests for Int class (256-bit arithmetic)
- test_bloom.cpp: 16 tests for Bloom filter operations
- test_bsgs_integration.cpp: 12 tests for BSGS/secp256k1
- test_framework.h: Lightweight assertion macros
- run_tests.cpp: Test runner with colored output
- Makefile: Added 'make test' target

Documentation:
- SECURITY.md: Security best practices and configuration

+++ src/secp256k1/SECP256k1.h
@@ -139,7 +139,8 @@ public:
 
 private:
 
-  uint8_t GetByte(char *str,int idx);
+  /* GetByte returns -1 on error instead of calling exit() */
+  int GetByte(char *str,int idx);
   Int GetY(Int x, bool isEven);
   Point GTable[256*32];       // Generator table
